move: tests to db packages
