+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/data/mapper/Data_01_Encoders_Decode.kt
@@ -1,4 +1,4 @@
-package test.entities
+package test.data.mapper
 
 import org.junit.Assert
 import org.junit.Test
